+++ pkg/services/plugins/client/client_test.go
